refactor: remove start.go
